+++ packages/robo/src/cli/utils/commands.ts
@@ -315,7 +315,9 @@ export async function registerCommands(
 	removedContextCommands: string[]
 ) {
 	const config = await loadConfig()
-	const { clientId, guildId, token } = env.discord
+	const clientId = env('discord.clientId')
+	const guildId = env('discord.guildId')
+	const token = env('discord.token')
 
 	if (!token || !clientId) {
 		logger.error(