+++ packages/robo/src/core/dotenv.ts
@@ -2,6 +2,7 @@ import { existsSync, readFileSync } from 'node:fs'
 import { readFile } from 'node:fs/promises'
 import path from 'node:path'
 import { IS_BUN } from '../cli/utils/runtime-utils.js'
+import { logger } from './logger.js'
 
 function parseEnvFile(envFileContent: string): { [key: string]: string } {
 	const lines = envFileContent.split('\n')
@@ -49,14 +50,29 @@ function parseEnvFile(envFileContent: string): { [key: string]: string } {
 	return newEnvVars
 }
 
-export async function loadEnv(options: { filePath?: string; sync?: boolean } = {}): Promise<void> {
+interface LoadEnvOptions {
+	filePath?: string
+	mode?: string
+	overwrite?: boolean
+	sync?: boolean
+}
+
+export async function loadEnv(options: LoadEnvOptions = {}): Promise<void> {
 	// No need to load .env file if using Bun (it's already loaded)
 	if (IS_BUN) {
 		return
 	}
 
-	const { filePath = path.join(process.cwd(), '.env') } = options
+	// Look for .env.{mode} file first, then fallback to standard .env
+	const { mode, overwrite } = options
+	let { filePath = path.join(process.cwd(), '.env') } = options
+
+	if (mode && existsSync(filePath + '.' + mode)) {
+		logger.debug('Found .env file for mode:', mode, ':', filePath + '.' + mode)
+		filePath = path.join(process.cwd(), '.env' + '.' + mode)
+	}
 	if (!existsSync(filePath)) {
+		logger.debug(`No .env file found at "${filePath}"`)
 		return
 	}
 
@@ -69,7 +85,10 @@ export async function loadEnv(options: { filePath?: string; sync?: boolean } = {
 
 	try {
 		for (const key in newEnvVars) {
-			if (key in envClone) continue // Don't overwrite existing values
+			// Don't overwrite existing values unless specified
+			if (!overwrite && key in envClone) {
+				continue
+			}
 
 			const visited = new Set<string>()
 			let value = newEnvVars[key]