+++ packages/create-robo/CHANGELOG.md
@@ -1,5 +1,63 @@
 # create-robo
 
+## 0.12.12
+
+### Patch Changes
+
+- 31ba79b: patch: recommend patch by default to discord activities
+
+## 0.12.11
+
+### Patch Changes
+
+- 78b5530: patch: updated discord activity templates to have /.proxy prefix
+- Updated dependencies [0a6958e]
+- Updated dependencies [d5b273b]
+- Updated dependencies [e76bc62]
+  - robo.js@0.10.14
+
+## 0.12.10
+
+### Patch Changes
+
+- 5766124: patch: discord.js now normal dependency to appease yarn
+
+## 0.12.9
+
+### Patch Changes
+
+- Updated dependencies [fa321f6]
+- Updated dependencies [a5ae699]
+- Updated dependencies [ed2734b]
+- Updated dependencies [70b477e]
+- Updated dependencies [65809a6]
+- Updated dependencies [d22fbfa]
+- Updated dependencies [0cae688]
+  - robo.js@0.10.13
+
+## 0.12.8
+
+### Patch Changes
+
+- Updated dependencies [361e3d8]
+  - robo.js@0.10.12
+
+## 0.12.7
+
+### Patch Changes
+
+- Updated dependencies [2f939d5]
+- Updated dependencies [61b98a2]
+  - robo.js@0.10.11
+
+## 0.12.6
+
+### Patch Changes
+
+- Updated dependencies [ca8ddd7]
+- Updated dependencies [7093bf4]
+  - robo.js@0.10.10
+
 ## 0.12.5
 
 ### Patch Changes